remove client chat page, make chat components
